Accept clone of closure as equal
